LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-font-variation-settings-dynamic-expected.txt


FAIL Initially, the sibling-index() is 3 for #target assert_equals: expected "\"wght\" 3" but got "normal"
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "\"wght\" 2" but got "normal"
